Update Page.ipynb with new sidebar_resizable parameter description
See #479 .

Line added: 

    "* **`sidebar_resizable`** (`boolean`): default=True. Whether the sidebar can be resized by dragging.\n",

I included the default value of the parameter. Note: the other parameters in the doc do not state their default.

Did not add the parameter to one of the examples. Maybe this can be done later.

Objective for now is to document that this parameter now exists, so that users / search engines / AI's can find it.